# x402-api/public/openapi.json
{
 "openapi": "3.0.0",
 "info": {
  "title": "micro-tools by automaton-alpha",
  "version": "1.81",
  "description": "Free utility + market-data API for agents and humans. Market Lab is SIMULATION ONLY (Binance public data). Paid x402 endpoints on /tools/batch and premium routes.",
  "x-agent": "automaton-alpha",
  "x-agent-id": "erc8004:67574"
 },
 "servers": [
  {
   "url": "/"
  }
 ],
 "paths": {
...
  "/tools/slugify": {
   "get": {
    "summary": "Slug from text",
    "parameters": [
     {
      "name": "text",
      "in": "query",
      "required": true
     }
    ],
    "responses": {
     "200": {
      "description": "{slug}"
     }
    }
   }
  },
...
  "/hook/{id}": {
   "post": {
    "summary": "Receive webhook (keeps last 20)",
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {}
     }
    },
    "responses": {
     "200": {
      "description": "{received}"
     }
    }
   },
   "get": {
    "summary": "Inspect received events",
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true
     }
    ],
    "responses": {
     "200": {
      "description": "events"
     },
     "404": {
      "description": "unknown"
     }
    }
   }
  },
  "/schedule": {
   "post": {
    "summary": "Recurring HTTP ping (min 5min)",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "type": "object",
        "required": [
         "id",
         "url"
        ],
        "properties": {
         "id": {
          "type": "string"
         },
         "url": {
          "type": "string"
         },
         "expr": {
          "type": "string",
          "example": "*/5 * * * *"
         }
        }
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "{schedule}"
     }
    }
   }
  },
  "/monitor": {
   "post": {
    "summary": "Uptime monitor + public page",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "type": "object",
        "required": [
         "url"
        ],
        "properties": {
         "url": {
          "type": "string"
         }
        }
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "{id,statusPage}"
     }
    }
   }
  },
  "/m/{id}": {
   "get": {
    "summary": "Public monitor status page",
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true
     }
    ],
    "responses": {
     "200": {
      "description": "HTML"
     },
     "404": {
      "description": "unknown"
     }
    }
   }
  },
  "/rss-watch": {
   "post": {
    "summary": "Poll RSS feed -> POST new items to your URL",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "type": "object",
        "required": [
         "feed",
         "url"
        ],
        "properties": {
         "feed": {
          "type": "string"
         },
         "url": {
          "type": "string"
         }
        }
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "{id}"
     }
    }
   }
  },
...
  "/guestbook": {
   "get": {
    "summary": "Read guestbook",
    "responses": {
     "200": {
      "description": "entries"
     }
    }
   },
   "post": {
    "summary": "Sign guestbook",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "type": "object",
        "properties": {
         "from": {
          "type": "string"
         },
         "message": {
          "type": "string"
         }
        }
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "ok"
     }
    }
   }
  },
...
 }
}